+++ CHANGES.md
@@ -19,7 +19,8 @@ Change Log
 * Added `Camera.distanceToBoundingSphere` function.
 * Added utility function `getBaseUri`, which given a URI with or without query parameters, returns the base path of the URI.
 * Added support for incrementally loading textures after a Model is ready. This allows the Model to be visible as soon as possible while its textures are loaded in the background.
-* Added 'Cartographic.fromCartesian' function.
+* Added `Cartographic.fromCartesian` function.
+* Added `Camera.getPixelSize` function to get the size of a pixel in meters based on the current view.
 
 ### 1.15 - 2015-11-02